🚚(frontend) move metadata to lib-components
As ResourceMetadata is used by several libs, we need to move it to a
lib that is build beforehand.
