+++ src/frontend/packages/lib_components/src/types/index.ts
@@ -3,11 +3,13 @@ export * from './api';
 export * from './apps/';
 export * from './file';
 export * from './jwt';
+export * from './metadata';
 export * from './models';
 export * from './Participant';
 export * from './PortabilityRequest';
 export * from './Organization';
 export * from './ResourceContext';
+export * from './RouteOptions';
 export * from './serviceWorker';
 export * from './tracks';
 export * from './User';
