+++ src/frontend/packages/lib_components/src/types/metadata.ts
@@ -1,0 +1,9 @@
+import { Resource, RouteOptions } from '.';
+
+export interface ResourceMetadata<T extends Resource> {
+  name: string;
+  description: string;
+  renders: string[];
+  parses: string[];
+  actions: RouteOptions<T>['actions'];
+}
