+++ src/frontend/packages/lib_video/src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './metadata';
-export * from './RouteOptions';
 export * from './SelectOptions';
 export * from './VideoPlayer';
